Update Voltage_divider.py
The problem with this code is with the loop. For every increase in limit, the number of iterations double accordingly. So for large values of limit, the number of iterations becomes limit times limit, and that takes a long time to execute. I need a better way to implement this.